refactor!: move structs & impls in middlewares.rs and routes.rs to more sensible places, apply formatter to all files
