+++ src/bin/scraper.rs
@@ -29,13 +29,7 @@ async fn main() -> Result<()> {
     tracing_subscriber::registry()
         .with(
             tracing_subscriber::EnvFilter::try_from_default_env()
-                .unwrap_or_else(|_| [
-                        "scraper=info",
-                        "scheduler=debug",
-                    ]
-                    .join(",")
-                    .into()
-                ),
+                .unwrap_or_else(|_| ["scraper=info", "scheduler=debug"].join(",").into()),
         )
         .with(tracing_subscriber::fmt::layer())
         .init();
