+++ src/lib.rs
@@ -1,4 +1,7 @@
+pub mod app;
+pub mod common;
 pub mod components;
+pub mod data;
 pub mod middlewares;
 pub mod routes;
 pub mod scraper;
